Finishing settings menu bottom section
moved the drawer code to a new file separate from the other screen which probably will be deleted at some point or repurposed.

Broke the code up a bit after getting it all working to refactor into smaller easier to understand chunks.
