+++ src/client/app/src/main/res/values/strings.xml
@@ -14,4 +14,7 @@
     <string name="filters">Filters</string>
     <string name="notifications">Notifications</string>
     <string name="saved">Saved</string>
+    <string name="about">About</string>
+    <string name="support">Support</string>
+    <string name="sign_out">Sign Out</string>
 </resources>
